refactor: 불필요 index 파일 제거 및 Utility Buttons 관련 컴포넌트 분리
