--- src/views/ProjectPage/components/common/index.ts
@@ -1,2 +1,0 @@
-export * from './EmptyContent';
-export * from './OvalSpinner';
