--- src/views/ProjectPage/components/filter/index.ts
@@ -1,1 +1,0 @@
-export * from './ProjectFilter';
