+++ src/views/ProjectPage/components/filter/project-filter-desktop.module.scss
@@ -79,26 +79,3 @@
     }
   }
 }
-
-.desktop-utility {
-  position: fixed;
-  right: 20px;
-  bottom: 30px;
-  border-radius: 52px;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
-  background: #242424;
-  padding: 16px 32px;
-  z-index: 999;
-  cursor: pointer;
-
-  & > span {
-    font-weight: 600;
-    font-size: 24px;
-    line-height: 48px;
-    color: #ffffff;
-  }
-
-  @include mobile-tablet {
-    display: none;
-  }
-}
