+++ src/views/ProjectPage/components/filter/project-filter-mobile.module.scss
@@ -1,35 +1,3 @@
-.utility {
-  position: fixed;
-  top: 80%;
-  right: 20px;
-  z-index: 100;
-
-  display: flex;
-  flex-direction: column;
-  gap: 10px 0;
-
-  .up {
-    width: 56px;
-    height: 56px;
-    border-radius: 50%;
-
-    background: #242424;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
-  }
-
-  .filter {
-    width: 56px;
-    height: 56px;
-    border-radius: 50%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    background: #ffffff;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
-  }
-}
-
 .modal {
   position: fixed;
   top: 0;
