+++ src/views/ProjectPage/components/index.ts
@@ -1,5 +1,4 @@
 export * from './common/EmptyContent';
-export * from './filter';
-export * from './project/ProjectCard';
-export * from './project/ProjectEnrollSection';
-export * from './project/ProjectList';
+export * from './common/OvalSpinner';
+export * from './filter/ProjectFilter';
+export * from './project';
