+++ src/views/ProjectPage/components/utility/utility.module.scss
@@ -1,0 +1,56 @@
+@import '@src/styles/mixins';
+
+.desktop-utility {
+  position: fixed;
+  right: 20px;
+  bottom: 30px;
+  border-radius: 52px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
+  background: #242424;
+  padding: 16px 32px;
+  z-index: 999;
+  cursor: pointer;
+
+  & > span {
+    font-weight: 600;
+    font-size: 24px;
+    line-height: 48px;
+    color: #ffffff;
+  }
+
+  @include mobile-tablet {
+    display: none;
+  }
+}
+
+.mobile-utility {
+  position: fixed;
+  top: 80%;
+  right: 20px;
+  z-index: 100;
+
+  display: flex;
+  flex-direction: column;
+  gap: 10px 0;
+
+  .up {
+    width: 56px;
+    height: 56px;
+    border-radius: 50%;
+
+    background: #242424;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
+  }
+
+  .filter {
+    width: 56px;
+    height: 56px;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background: #ffffff;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
+  }
+}
